backend/requirements.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==3.0.8
django-rest-auth==0.9.5
djangorestframework==3.11.1
environ==1.0
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
